Trickwine Touchups! (#261)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds and spellchecks status effect descs and fluff text to trickwine
buffs and debuffs where it was absent (which was like 80% of cases for
some reason). Also removes the previously copious use of the word
"seems" in said fluff text, because my english teachers always said
"seems" was a lousy way to describe things. If any of these need further
edits I'd be happy to rewrite them again.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Having more player-facing flavorful feedback is good, especially on
stuff where it can be difficult to tell what, if anything, is happening
without it. Also the mention of them in the CL might encourage at least
one guy to try trickwines, and nobody's used them since forever. (and
typo fixing is always good anyhow)
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: The Galaxy's Sole Trickwine User
add: Effect text to trickwines missing it. Go forth in the name of The
Ashen Huntsman, secure in the knowledge that you actually know what
these things are doing.
spellcheck: fixed a few typos in existing trickwine effect texts.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Ossa88 (SYNAPSE) <MegaMan101@gmail.com>