Better Equip Delay Modularization (#274)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Removed the quick and dirty define overrides in place of redefining the
proc

Currently the equip/unequip proc will only trigger if the item is a suit
(as it would before).


Also discovered that neck items were never given a proper equip/unequip
so they have always been an instant equip/unequip/strip.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

The old Define override causes warnings.
Also because whenever you stripped others the proc would multiply the
set time by a value. So 0 * x meant that it was always 0.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
balance: Stripping others has had its values corrected
fix: Equip Delay Defines no longer complain on Github and in tests.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->