[Bug Fix] Aegis TEG corrected (#268)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Corrects the Aegis's TEG.

Seems that a layer 4 pipe was used on the exhaust part of the TEG which
prevented it from working.

I had heard people were having issues but every time this came up I was
at work and I never knew what the ship was that had acted up when I was
told. I only just noticed #170

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<details>
  <summary>Pictures</summary>


![StrongDMM_WxnJoC8kcf](https://github.com/user-attachments/assets/cf75d052-7da3-4dbc-abff-a7dedfd2a029)

![StrongDMM_Q2bP4R8v0S](https://github.com/user-attachments/assets/09c564c7-0098-4412-bcb6-fae1a718464e)

</details>


## Why It's Good For The Game

Fixes the Aegis's TEG.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: Aegis TEG Piping has been corrected.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->