Ports the monkestation ethereal sprites. (#265)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Ports the monkestation ethereal sprites from monkestation to pentest. 


![image](https://github.com/user-attachments/assets/5158dab7-49f5-4c05-8cec-4ccef978faf2)
Permission granted by @aberrantQuesrist for the art usage.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

From what I can tell the server is trying to move to an altered version
of TG lore instead of shiptest lore.


## Changelog

:cl:
imageadd: Ported the monkestation ethereal sprites.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Ossa88 (SYNAPSE) <MegaMan101@gmail.com>