Adjusts the Placement of Teleporter Pieces (#281)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Teleporters worked on the following principle: Console (controls it) -->
Power Station (turns it on and off) --> Hub (the gate portion), placed
in that order, cardinally.

I thought this was too limiting, and the code for this is ancient, so I
refactored it somewhat to allow more loose formations. Specifically the
Hub can now be in the middle of the Power Station and Console. Power
Station --> Hub <-- Console. The Hub and Power Station still have to be
next to each other. You can still use the original configuration too, or
various L-type designs. This should also prevent weird recursions in
links.

I have also cleaned up the relevant code, added some exceptions for when
things are destroyed to fix some weird visual effects and such.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

More of a QoL. I think having more design space is a good thing. And the
Hub being in the middle just looks right.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
refactor: Teleporter pieces can now be arranged to have the hub in the
middle.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->