Beach fishing dock (#262)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds a new ENTRY level ruin, with some mild loot for self sufficiency,
and some nice items for roleplay. It has some salvageable solars and
some other equipment. This is a isolated fishing dock that was overrun
by a carp migration.

This is a redone pull request because I did an oopsie daisy.

It has several fishing related items, including lures, line, and a
fishing portal generator, alongside a single guarded revolver.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

It's meant for basic ships and newer players to have the option to make
money via fishing/mining, or get a renewable source of energy for their
ship, ideally.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->
<details>
<summary>Ruin Screenshot</summary>
<br>

![StrongDMM-2025-04-11 22 44
04](https://github.com/user-attachments/assets/6f989237-4976-4dc7-82a6-5bfdd4348532)

</details>



## Changelog

:cl:
add: Added a new low threat beach ruin.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Ossa88 (SYNAPSE) <MegaMan101@gmail.com>