Adds soggy woggies (God Bless America, Roll Tide, and Praise King Saban) (#189)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Adds and ~~avenges~~ betrays pissy shitties, the beloved movie theater
snack of every American childhood. (NOW CALLED SOGGY WOGGIES DUE TO
PROTESTS FROM THE SOUTHERN HEMISPHERE.) Finally, you can pour soda into
your popcorn and enjoy it (almost) the right way, as a chunky brown
slurry of salty-sweet goodness. Eagles will cry and Lady Liberty will
tip her crown to you as you indulge in the snack food version of bravery
and freedom.

Allows popcorn to accept any drink reagents, and adds a recipe for the
above.


![pissyshitties](https://github.com/user-attachments/assets/ba79e656-08d3-46bf-ae27-0efb667ee0a6)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

~~Big Spessmen has kept us down for far too long. They say the name is
"gross" and the food is "disgusting" and "not real". How long are we
going to stand for this mistreatment? Is it a crime to love mixing soda
with popcorn?~~ SOGGY WOGGIES, now 10% less gross in name.

Popcorn is more customizable... and it can also be poisoned.

The slurry code also has future applicability for adding drink reagents
to other food items, if desired. (OTHER TYPES OF SLURRIES. BEER
CASSEROLE. METH PIE.)

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: soggy woggies snack item and trash
add: drinks can now be added to popcorn
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: vect0r <71346830+Vect0r2@users.noreply.github.com>
Co-authored-by: Ossa88 (SYNAPSE) <MegaMan101@gmail.com>